refactor(scraping): move scrapers to separate file
